--- 365Days-Website/src/App.jsx
@@ -1,13 +1,0 @@
-import './App.css'
-
-function App() {
- 
-
-  return (
-    <>
-      <h1 className='font-bold text-3xl bg-slate-400 rounded-lg mx-auto p-2'> 365 Coding Days Tech Community</h1>
-    </>
-  )
-}
-
-export default App
