--- 365Days-Website/src/components/navbar.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const navbar = () => {
-  return (
-    <div>navbar</div>
-  )
-}
-
-export default navbar
