+++ src/App.jsx
@@ -1,0 +1,13 @@
+import "./App.css";
+
+function App() {
+	return (
+		<>
+			<h1 className="font-bold text-3xl bg-slate-400 rounded-lg mx-auto p-2">
+				365 Days Coding Tech Community
+			</h1>
+		</>
+	);
+}
+
+export default App;
