+++ src/components/navbar.jsx
@@ -1,0 +1,5 @@
+const navbar = () => {
+	return <div>navbar</div>;
+};
+
+export default navbar;
